refactor: split utils into focused sub-packages (rand, ip, packet)
